binotel-ai-bot/requirements.txt
# requirements.txt
Flask
requests
python-dotenv
openai-whisper

